fix prompt being shown twice
Reason (as I can see it):
The interpolation of the request params is done twice.
Both interpolations were called using the context RENDER_PURPOSE_RENDER which triggers
the prompt in `local-template-tags.ts`.
This file contains the "prompt" function as a plugin.
`context.app.prompt` is called there without checking for the already set cached value when context === "send".

This "fix" renders the first time with renderPurpose set to "no-render" so no prompt is called in this case since no
plugins (the "prompt") are called anyway.